+++ packages/client/src/types.ts
@@ -168,6 +168,12 @@ export type PublishOptions = {
    */
   forceCodec?: PreferredCodec;
   /**
+   * When using a preferred codec, force the use of a single codec.
+   * Enabling this, it will remove all other supported codecs from the SDP.
+   * Defaults to false.
+   */
+  forceSingleCodec?: boolean;
+  /**
    * The preferred scalability to use when publishing the video stream.
    * Applicable only for SVC codecs.
    */