+++ sample-apps/react-native/dogfood/types.ts
@@ -162,20 +162,26 @@ export type ScreenTypes =
   | 'active-call'
   | 'loading';
 
-export type LocalAttachmentType = Record<string, unknown>;
-export type LocalChannelType = Record<string, unknown>;
-export type LocalCommandType = string;
-export type LocalEventType = Record<string, unknown>;
-export type LocalMessageType = Record<string, unknown>;
-export type LocalReactionType = Record<string, unknown>;
-export type LocalUserType = Record<string, unknown>;
+type LocalAttachmentType = Record<string, unknown>;
+type LocalChannelType = Record<string, unknown>;
+type LocalCommandType = string;
+type LocalEventType = Record<string, unknown>;
+type LocalMessageType = Record<string, unknown>;
+type LocalReactionType = Record<string, unknown>;
+type LocalUserType = Record<string, unknown>;
+type LocalPollOptionType = Record<string, unknown>;
+type LocalPollType = Record<string, unknown>;
+type LocalMemberType = Record<string, unknown>;
 
 export type StreamChatGenerics = {
   attachmentType: LocalAttachmentType;
   channelType: LocalChannelType;
   commandType: LocalCommandType;
   eventType: LocalEventType;
+  memberType: LocalMemberType;
   messageType: LocalMessageType;
+  pollOptionType: LocalPollOptionType;
+  pollType: LocalPollType;
   reactionType: LocalReactionType;
   userType: LocalUserType;
 };