+++ packages/react-sdk/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.7.19](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.7.18...@stream-io/video-react-sdk-1.7.19) (2024-11-21)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `1.11.4`
+* `@stream-io/video-react-bindings` updated to version `1.1.21`
 ## [1.7.18](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.7.17...@stream-io/video-react-sdk-1.7.18) (2024-11-20)
 
 ### Dependency Updates